refactor(application): split run_l1 and run_command helpers
- Extract L1 prepare, evaluator, and evidence helpers beside thin orchestration
- Sink L0 evaluator worker and environment prepare out of run_command
- Retarget ACP readiness monkeypatch and architecture gates (Refs: #31)
